requirements.txt
tweepy>=4.14.0
anthropic>=0.34.0
python-dotenv>=1.0.0
yfinance>=0.2.40
APScheduler>=3.10.4
pytz>=2024.1
requests>=2.31.0
